Moved GLib testcases
The GLib testcases are now in the package that corresponds to the functionality under test (io.github.jwharm.javagi.interop).
