+++ modules/glib/src/test/java/io/github/jwharm/javagi/interop/ArrayTest.java
@@ -1,6 +1,5 @@
-package io.github.jwharm.javagi.test.glib;
+package io.github.jwharm.javagi.interop;
 
-import io.github.jwharm.javagi.interop.Interop;
 import org.gnome.glib.GLib;
 import org.gnome.glib.HashTable;
 import org.junit.jupiter.api.Test;
